Improved pathfinding speed
Changed the pathfinding system a little, the game will now spawn a tile on every walkable tile at the start of a level, and then during pathfinding will simply modulate them.

This improvement has made the pathfinding so fast that the system can find a path of up to a 50 blocks radius in under 10ms.