Map Editor Update
Lots of new smaller functionality added to the map editor. Will take a break for a while to focus on the procedural branch for now.

Added the ability to work on multiple layers, map shows you where you're placing tiles on each layer.

Will need to add the ability to rotate the map in the future, but otherwise progress is great so far.